+++ src/main/java/org/mattlang/jc/engine/evaluation/parameval/MaterialDescription.java
@@ -1,5 +1,7 @@
 package org.mattlang.jc.engine.evaluation.parameval;
 
+import static java.util.Objects.requireNonNull;
+
 import org.mattlang.jc.material.Material;
 
 import lombok.Getter;
@@ -8,8 +10,11 @@ import lombok.Getter;
  * A Description of Material used in Configurations.
  * It specifies Pieces and Pawns.
  *
- * Pieces must always fix specified.
+ * Pieces must be fix specified but can be combined with + or *.
  * Pawns can be omitted
+ *
+ * + == more than the defined material
+ * * == more or equal than the defined material
  *
  * Example
  *
@@ -18,6 +23,12 @@ import lombok.Getter;
  * RRB   == two Rooks, a Bishop, any number of Pawns since they are not specified
  *
  * RQ 2P == a rook, a queen and two pawns
+ *
+ * RQ+ == a rook, a queen and some more unspecified material
+ * RB* == a rook, a bishop at least and maybe even more unspecified material
+ *
+ * X == any material. This makes only sense in certain situations like describing a weaker material side when a stronger
+ * material description is given which would mean that it is anything but at least weaker then the stronger side
  */
 @Getter
 public class MaterialDescription {
@@ -29,22 +40,22 @@ public class MaterialDescription {
 
     private boolean pawnsUnspecific = false;
 
-    /* complete unspecified. this makes only sense for the weaker side: means it is unspecified, but anyway lower/weaker than the stronger side. */
-    private boolean unspecified = false;
+    private MaterialComparison comparison = MaterialComparison.EQUAL;
 
-    public MaterialDescription(boolean unspecified) {
-        this.unspecified = unspecified;
+    public MaterialDescription(MaterialComparison comparison) {
+        this.comparison = requireNonNull(comparison);
     }
 
-    public MaterialDescription(Material material, boolean pawnsUnspecific) {
-        this.material = material;
-        this.pawnsUnspecific = pawnsUnspecific;
+    public MaterialDescription(Material material, boolean pawnsUnspecific, MaterialComparison comparison) {
+        this.material = requireNonNull(material);
+        this.pawnsUnspecific = requireNonNull(pawnsUnspecific);
+        this.comparison = requireNonNull(comparison);
     }
 
     public static MaterialDescription parseDescr(String descr) {
         descr = descr.trim();
         if ("X".equals(descr)) {
-            return new MaterialDescription(true);
+            return new MaterialDescription(MaterialComparison.UNSPECIFIED);
         }
         String[] splitted = descr.split(" ");
         String pieces = splitted[0];
@@ -59,7 +70,18 @@ public class MaterialDescription {
             matDescr += convertPawnStr(pawnstr);
         }
 
-        return new MaterialDescription(new Material(matDescr), pawnsUnspecified);
+        MaterialComparison comparison = MaterialComparison.EQUAL;
+
+        if (matDescr.contains("+")) {
+            comparison = MaterialComparison.MORE;
+        }
+        if (matDescr.contains("*")) {
+            comparison = MaterialComparison.MORE_OR_EQUAL;
+        }
+        matDescr = matDescr.replace("+", "");
+        matDescr = matDescr.replace("*", "");
+
+        return new MaterialDescription(new Material(matDescr), pawnsUnspecified, comparison);
     }
 
     private static String convertPawnStr(String pawnstr) {
@@ -86,9 +108,27 @@ public class MaterialDescription {
      * @return
      */
     public boolean matches(Material other) {
-        if (unspecified) {
+        switch (comparison) {
+        case UNSPECIFIED:
             return true;
+        case EQUAL:
+            return compareEqual(other);
+        case MORE:
+            return compareMore(other);
+        case MORE_OR_EQUAL:
+            return compareEqual(other) || compareMore(other);
+        default:
+            throw new IllegalStateException("unkown Comparison value: " + comparison);
         }
+
+    }
+
+    private boolean compareMore(Material other) {
+        return other.hasMoreWhiteMat(material);
+    }
+
+    private boolean compareEqual(Material other) {
+
         if (material.getPieceMat() != other.getPieceMat()) {
             return false;
         }