+++ src/main/java/org/mattlang/jc/engine/evaluation/parameval/endgame/EndgameFunction.java
@@ -1,8 +1,9 @@
 package org.mattlang.jc.engine.evaluation.parameval.endgame;
 
 import org.mattlang.jc.board.BoardRepresentation;
+import org.mattlang.jc.engine.evaluation.parameval.ParameterizedMaterialEvaluation;
 
 public interface EndgameFunction {
 
-    int evaluate(BoardRepresentation board, int stronger, int weaker, int materialEval);
+    int evaluate(BoardRepresentation board, int stronger, int weaker, ParameterizedMaterialEvaluation matEvaluation);
 }